Fix before scenario running multiple times
