+++ src/Sylius/Bundle/CoreBundle/Behat/HookContext.php
@@ -1,0 +1,76 @@
+<?php
+/*
+ * This file is part of the Sylius package.
+ *
+ * (c) Paweł Jędrzejewski
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Sylius\Bundle\CoreBundle\Behat;
+
+use Behat\Behat\Context\Context;
+use Behat\Behat\Hook\Scope\BeforeScenarioScope;
+use Behat\Symfony2Extension\Context\KernelAwareContext;
+use Doctrine\Common\DataFixtures\Purger\ORMPurger;
+use Symfony\Component\DependencyInjection\ContainerInterface;
+use Symfony\Component\HttpKernel\KernelInterface;
+
+/**
+ * @author Gonzalo Vilaseca <gvilaseca@reiss.co.uk>
+ */
+class HookContext implements Context, KernelAwareContext
+{
+    /**
+     * @var KernelInterface
+     */
+    protected $kernel;
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setKernel(KernelInterface $kernel)
+    {
+        $this->kernel = $kernel;
+    }
+
+    /**
+     * @BeforeScenario
+     */
+    public function purgeDatabase(BeforeScenarioScope $scope)
+    {
+        $entityManager = $this->getService('doctrine.orm.entity_manager');
+        $entityManager->getConnection()->getConfiguration()->setSQLLogger(null);
+
+        $entityManager->getConnection()->executeUpdate("SET foreign_key_checks = 0;");
+
+        $purger = new ORMPurger($entityManager);
+        $purger->purge();
+
+        $entityManager->getConnection()->executeUpdate("SET foreign_key_checks = 1;");
+        $entityManager->clear();
+    }
+
+    /**
+     * Get service by id.
+     *
+     * @param string $id
+     *
+     * @return object
+     */
+    protected function getService($id)
+    {
+        return $this->getContainer()->get($id);
+    }
+
+    /**
+     * Returns Container instance.
+     *
+     * @return ContainerInterface
+     */
+    protected function getContainer()
+    {
+        return $this->kernel->getContainer();
+    }
+}
